Some fixed

+++ app/Services/ContentLoaders/LocalContentLoader.php
@@ -14,6 +14,7 @@ use App\Services\Facades\LogRec;
 use Illuminate\Http\Request;
 use App\Models\ContentItem;
 use App\Services\Facades\ContentLoader;
+use Illuminate\Http\File;
 
 final class LocalContentLoader extends ContentLoader
 {
@@ -32,8 +33,15 @@ final class LocalContentLoader extends ContentLoader
         $contentItem->fill($request->get('contentItem'));
 
         // Upload content item
-        if ($request->input('contentItem')['download']['link'] && $request->input('contentItem')['type'] !== 'reference') {
-            if (Storage::disk('temp')->exists(basename($request->input('contentItem')['download']['link']))) {
+        if ($request->input('contentItem')['download']['link'] && $request->input('contentItem')['type'] !== 'reference' && Storage::disk('temp')->exists(basename($request->input('contentItem')['download']['link']))) {
+            // If type is set to cloud upload
+            if($request->input('contentItem')['type'] === 'cloud') {
+                $contentItem->download = $this->moveItemToCloud(
+                    $request->input('contentItem')['download']['link'],
+                    'content-item_' . $contentItem->id . '_' . time(),
+                    'content-items');
+            } else {
+                // Upload file locally
                 $this->deleteOldContentFile(basename($contentItem->download['link']));
                 $contentItem->download = $this->moveToContentFolder($request, $contentItem->id);
             }
@@ -48,7 +56,6 @@ final class LocalContentLoader extends ContentLoader
         }
 
         $contentItem->save();
-
         return $contentItem;
     }
 
@@ -75,7 +82,6 @@ final class LocalContentLoader extends ContentLoader
         }
 
         $contentItem->save();
-        $this->deleteTempFolder();
         return $contentItem;
     }
 
@@ -131,14 +137,6 @@ final class LocalContentLoader extends ContentLoader
         ];
     }
 
-    private function deleteTempFolder()
-    {
-        $chance = rand(0, 10);
-        if ($chance === 1) {
-            Storage::disk('local')->deleteDirectory('\public\temp');
-        }
-    }
-
     private function deleteOldImageFile($imageFile)
     {
         Storage::disk('images')->delete($imageFile);
@@ -147,6 +145,18 @@ final class LocalContentLoader extends ContentLoader
     private function deleteOldContentFile($contentFile)
     {
         Storage::disk('content')->delete($contentFile);
+    }
+
+    private function moveItemToCloud($tempInputLink, $newFileName, $storage)
+    {
+        $tempFile = basename($tempInputLink);
+        $fileExtension = pathinfo($tempFile, PATHINFO_EXTENSION);
+        $newFile = $newFileName . '.' . $fileExtension;
+
+        $fullFilePath = Storage::disk('temp')->getDriver()->getAdapter()->getPathPrefix() . "/" . $tempFile;
+        $file = new File($fullFilePath);
+        Storage::disk('spaces')->putFileAs($storage, $file, $newFile);
+        return ['link' => $storage . '/' . $newFile];
     }
     #endregion
 }
